adds doc strings to the ke_data loaders
