--- src/transformerkp/data/extraction/loader.py
@@ -1,832 +1,0 @@
-import logging
-import pathlib
-from collections import defaultdict
-from typing import Union, Dict, List
-
-from datasets import get_dataset_split_names
-from datasets import Dataset
-from datasets import DatasetDict
-from datasets import IterableDataset
-from datasets import IterableDatasetDict
-from datasets import load_dataset
-
-from transformerkp.data.base import KPDataset
-from transformerkp.data.extraction.args import KEDataArguments
-from transformerkp.data.extraction import args
-
-logger = logging.getLogger(__name__)
-
-
-class KEDataset(KPDataset):
-
-    _label_to_id: Dict = {"B": 0, "I": 1, "O": 2}
-    _id_to_label: Dict = {0: "B", 1: "I", 2: "O"}
-    _num_labels: int = 3
-
-    def __init__(self, data_args: KEDataArguments):
-        
-        super().__init__()
-        self.data_args: KEDataArguments = data_args
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = None
-        self._validation: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = None
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = None
-        # TODO: needs to be removed. No need of handling it over here
-        # self._tokenizer: Any = AutoTokenizer.from_pretrained(data_args.tokenizer, use_fast=True, add_prefix_space=True)
-        self._splits_to_load: Union[List[str], None] = self.data_args.splits
-        self._text_column_name: Union[str, None] = (
-            self.data_args.text_column_name if self.data_args is not None else None
-        )
-        self._label_column_name: Union[str, None] = (
-            self.data_args.label_column_name if self.data_args is not None else None
-        )
-        # TODO: will remove after proper verificiation
-        # if self._data_args.max_seq_length is None:
-        #     self._data_args.max_seq_length = self._tokenizer.model_max_length
-        # if self._data_args.max_seq_length > self._tokenizer.model_max_length:
-        #     logger.warning(
-        #         f"The max_seq_length passed ({self._data_args.max_seq_length}) is larger than the maximum length for the"
-        #         f"model ({self._tokenizer.model_max_length}). Using max_seq_length={self._tokenizer.model_max_length}."
-        #     )
-        # self._max_seq_length: int = min(
-        #     self._data_args.max_seq_length, self._tokenizer.model_max_length
-        # )
-        self._padding: Union[str, bool] = (
-            "max_length" if self.data_args.pad_to_max_length else False
-        )
-        self.__preprocess_function = self.data_args.preprocess_func
-        self._datasets: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = None
-        self.__load_kp_datasets()
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        """Gets the train split"""
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        """Gets the validation split"""
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        """Gets the test split"""
-        return self._test
-
-    @property
-    def text_column_name(self) -> Union[str, None]:
-        """Get the column name of the input text from which keyphrases needs to be extracted"""
-        return self._text_column_name
-
-    @property
-    def label_column_name(self) -> Union[str, None]:
-        """Get the column name of the column which contains the BIO labels of the tokens"""
-        return self._label_column_name
-
-    @property
-    def label_to_id(self) -> Dict:
-        """Get the label to id mapping used for creating the dataset"""
-        return self._label_to_id
-
-    @property
-    def padding(self) -> Union[str, bool]:
-        """Get the padding strategy to be used for preprocessing the dataset for training and evaluation"""
-        return self._padding
-
-    @property
-    def label_all_tokens(self) -> bool:
-        """Get the label_all_tokens property which decides whether to put the label for one word on all
-        sub-words generated by that word or just on the one (in which case the other tokens will have a padding index).
-        This will be later used while preprocessing the dataset for training and evaluation
-        """
-        return self._data_args.label_all_tokens
-
-    @property
-    def max_seq_length(self) -> int:
-        """Gets the max sequence length to be used while preprocessing the dataset for training and evaluation"""
-        return self._max_seq_length
-
-    def __load_kp_datasets(self) -> None:
-        """Loads the training, validation and test splits from either an existing dataset from Huggingface hub or from provided files.
-
-        """
-        if self.data_args.dataset_name is not None:
-            dataset_splits = get_dataset_split_names(self.data_args.dataset_name, "extraction")
-            self._splits_to_load = list(set(dataset_splits).intersection(set(self.data_args.splits)))
-            logger.info(f"Only loading the following splits {self._splits_to_load}")
-            # Downloading and loading a dataset from the hub.
-            self._datasets = load_dataset(
-                self.data_args.dataset_name,
-                self.data_args.dataset_config_name,
-                split=self._splits_to_load,
-                cache_dir=self.data_args.cache_dir,
-            )
-        else:
-            # TODO: What if the train, validation and test files are in different formats - we cannot allow it.
-            data_files = {}
-            if self.data_args.train_file is not None:
-                data_files["train"] = self.data_args.train_file
-                extension = pathlib.Path(self.data_args.train_file).suffix.replace(".", "")
-                logger.info(f"Loaded training data from {self.data_args.train_file}")
-            if self.data_args.validation_file is not None:
-                data_files["validation"] = self.data_args.validation_file
-                extension = pathlib.Path(self.data_args.validation_file).suffix.replace(".", "")
-                logger.info(f"Loaded validation data from {self.data_args.validation_file}")
-            if self.data_args.test_file is not None:
-                data_files["test"] = self.data_args.test_file
-                extension = pathlib.Path(self.data_args.test_file).suffix.replace(".", "")
-                logger.info(f"Loaded test data from {self.data_args.test_file}")
-
-            logger.info(f"Only loading the following splits {self._splits_to_load}")
-            self._datasets = load_dataset(
-                extension, 
-                data_files=data_files, 
-                cache_dir=self.data_args.cache_dir,
-                split=self._splits_to_load,
-            )
-
-        if self.__preprocess_function:
-            if self._datasets:
-                self._datasets = self._datasets.map(
-                    self.__preprocess_function,
-                    num_proc=self.data_args.preprocessing_num_workers,
-                )
-                logger.info(f"preprocessing done with the provided customized preprocessing function")
-        
-        self._datasets = self.__create_dataset_dict_from_splits(self._datasets)
-
-        if "train" in self._datasets:
-            column_names = self._datasets["train"].column_names
-        elif "validation" in self._datasets:
-            column_names = self._datasets["validation"].column_names
-        elif "test" in self._datasets:
-            column_names = self._datasets["test"].column_names
-        else:
-            raise AssertionError(
-                "neither train, validation or test dataset is available"
-            )
-
-        if self._text_column_name is None:
-            self._text_column_name = (
-                # TODO: convey this information properly in the documentation
-                "document" if "document" in column_names else column_names[1]
-            )  # either document or 2nd column as text i/p
-
-        assert self._text_column_name in column_names
-
-        if self._label_column_name is None:
-            self._label_column_name = (
-                "doc_bio_tags" if "doc_bio_tags" in column_names else None
-            )
-            # TODO: convey this information properly in the documentation
-            if len(column_names) > 2:
-                self._label_column_name = column_names[2]
-
-        if self._label_column_name is not None:
-            assert self._label_column_name in column_names
-
-        # TODO: to be done outside before training
-        # self._datasets = tokenize_and_align_labels(
-        #     datasets=self._datasets,
-        #     text_column_name=self._text_column_name,
-        #     label_column_name=self._label_column_name,
-        #     tokenizer=self._tokenizer,
-        #     label_to_id=self._label_to_id,
-        #     padding=self._padding,
-        #     label_all_tokens=self._data_args.label_all_tokens,
-        #     max_seq_len=self._max_seq_length,
-        #     num_workers=self._data_args.preprocessing_num_workers,
-        #     overwrite_cache=self._data_args.overwrite_cache,
-        # )
-
-        if "train" in self._datasets:
-            self._train = self._datasets["train"]
-        if "validation" in self._datasets:
-            self._validation = self._datasets["validation"]
-        if "test" in self._datasets:
-            self._test = self._datasets["test"]
-
-    def __create_dataset_dict_from_splits(self, data_splits: List[Dataset]) -> DatasetDict:
-        data_dict = defaultdict(Dataset)
-        for split_name, data_split in zip(self._splits_to_load, data_splits):
-            data_dict[split_name] = data_split
-        return DatasetDict(data_dict)
-
-
-class InspecKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "validation", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir = None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.InspecKEDataArguments = args.InspecKEDataArguments()
-        self._data_args.splits = splits
-        self._data_args.max_seq_length = max_seq_length
-        self._data_args.label_all_tokens = label_all_tokens
-        self._data_args.cache_dir = cache_dir
-        self._dataset: KEDataset  = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-class NUSKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.NUSKEDataArguments = args.NUSKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-class KDDKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.KDDKEDataArguments = args.KDDKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-class KrapivinKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.KrapivinKEDataArguments = args.KrapivinKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class KP20KKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "validation", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.KP20KKEDataArguments = args.KP20KKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset  = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-class WWWKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.WWWKEDataArguments = args.WWWKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class LDKP3KSmallKEDataset(KEDataset):
-    """"""
-    def __init__(self):
-        super().__init__(data_args=args.LDKP3KSmallKEDataArguments())
-
-class LDKP3KMediumKEDataset(KEDataset):
-    """"""
-    def __init__(self):
-        super().__init__(data_args=args.LDKP3KMediumKEDataArguments())
-
-class LDKP3KLargeKEDataset(KEDataset):
-    """"""
-    def __init__(self):
-        super().__init__(data_args=args.LDKP3KLargeKEDataArguments())
-
-
-class LDKP10KSmallKEDataset(KEDataset):
-    """"""
-    def __init__(self):
-        super().__init__(data_args=args.LDKP10KSmallKEDataArguments())
-
-
-class LDKP10KMediumKEDataset(KEDataset):
-    """"""
-    def __init__(self):
-        super().__init__(data_args=args.LDKP10KMediumKEDataArguments())
-
-
-class LDKP10KLargeKEDataset(KEDataset):
-    """"""
-    def __init__(self):
-        super().__init__(data_args=args.LDKP10KLargeKEDataArguments())
-
-
-class KPTimesKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "validation", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.KPTimesKEDataArguments = args.KPTimesKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset  = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class OpenKPKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "validation", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.OpenKPKEDataArguments = args.OpenKPKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset  = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class SemEval2010KEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.SemEval2010KEDataArguments = args.SemEval2010KEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset  = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class SemEval2017KEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "validation", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.SemEval2017DataArguments = args.SemEval2017KEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset  = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class KPCrowdKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.KPCrowdKEDataArguments = args.KPCrowdKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class DUC2001KEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.KPCrowdKEDataArguments = args.DUC2001KEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-
-class CSTRKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["train", "test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.CSTRKEDataArguments = args.CSTRKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class PubMedKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.PubMedKEDataArguments = args.PubMedKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
-
-class CiteulikeKEDataset(KPDataset):
-    """"""
-    def __init__(
-            self,
-            splits: list = ["test"],
-            max_seq_length: int = 512,
-            label_all_tokens: bool = True,
-            cache_dir=None,
-    ):
-        """
-
-        Args:
-            splits:
-            max_seq_length:
-            label_all_tokens:
-            cache_dir:
-        """
-        super().__init__()
-        self._data_args: args.CiteulikeKEDataArguments = args.CiteulikeKEDataArguments()
-        self._data_args.splits = splits
-        self._dataset: KEDataset = KEDataset(self._data_args)
-        self._train: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.train
-        self._validation: Union[
-            DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.validation
-        self._test: Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None] = self._dataset.test
-
-    @property
-    def train(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._train
-
-    @property
-    def validation(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._validation
-
-    @property
-    def test(self) -> Union[DatasetDict, Dataset, IterableDatasetDict, IterableDataset, None]:
-        return self._test
-
